Add a Packet_v2 class and refactor the packet.py module
