+++ docs/api/larpix.rst
@@ -5,7 +5,15 @@ LArPix Control Core
 .. automodule:: larpix.chip
 .. automodule:: larpix.key
 .. automodule:: larpix.configuration
-.. automodule:: larpix.packet
+
+Packets
+^^^^^^^
+
+.. automodule:: larpix.packet.packet_v1
+.. automodule:: larpix.packet.packet_v2
+.. automodule:: larpix.packet.timestamp_packet
+.. automodule:: larpix.packet.message_packet
+.. automodule:: larpix.packet.packet_collection
 
 Configuration registers
 ^^^^^^^^^^^^^^^^^^^^^^^
