+++ larpix/packet/__init__.py
@@ -1,0 +1,7 @@
+from .packet_v1 import *
+from .packet_v2 import *
+from .timestamp_packet import *
+from .message_packet import *
+Packet = Packet_v1
+
+from .packet_collection import *
